Decorating references in a separate thread
- to unblock message receiving thread, decorating references (by adding `lineText` attribute) in a separate thread
